fix(codec): f32-vs-integer limit checks aliased with saturation
`T::MAX as f32` rounds *up* to `T::MAX + 1` for a target with >= 24
significant bits, so comparing an f32 against it cannot distinguish an
in-range value from one the saturating `as` cast clamped. Three sites
shared the mistake:

- `scx-codec/src/cast.rs` — the read-side fail-loud cast gate returned
  `i32::MAX` for an input of 2^31 (likewise u32/2^32, i64/2^63) rather
  than rejecting it. Reachable from
  `to_anndata(data_dtype="int32", allow_lossy=False)` on a float-encoded
  shard: a gate whose whole purpose is to reject this answered with a
  wrong number.
- `scx-codec/src/{dispatch,value_encoding}.rs` — the write side encoded
  2^32 as 4294967295 on disk, and hard-failed `from_anndata` on any
  integer-valued float layer above u32::MAX (e.g. 1e10) that the f64
  detector already handled by falling back to Float32.
- `scx-ops/src/helpers.rs` — the compact/merge re-encode path carries its
  own copy of the same u32 check.

Bounds are now the true exclusive limits, built as powers of two (exact
in f32 at every width) and hoisted to `const`. Integer range checks are
written as `!range.contains(&v)`, which also rejects NaN: previously
`scx_ops::encode_value` silently wrote NaN to disk as 0, while the codec's
`encode_f32` already rejected it.

`OpsError::ValueOutOfRange.max` widens f32 -> f64 so the reported uint32
limit is representable in the field that reports it.

+++ scx-ops/src/error.rs
@@ -63,11 +63,13 @@ pub enum OpsError {
     #[error("CSR shape mismatch on append: {detail}")]
     ShapeMismatch { detail: String },
 
+    /// `max` is `f64`, not `f32`: `u32::MAX as f32` rounds up to 2³², so an
+    /// `f32` limit field would report a bound one larger than the one enforced.
     #[error("f32 value {value} out of range for {encoding} encoding (max {max})")]
     ValueOutOfRange {
         value: f32,
         encoding: &'static str,
-        max: f32,
+        max: f64,
     },
 
     #[error(transparent)]
